requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arandr==0.1.10
astroid==2.5.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
automationhat==0.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==7.1.2
colorama==0.4.4
colorzero==1.1
cryptography==3.3.2
cupshelpers==1.0
dataclasses-json==0.5.7
dbus-python==1.2.16
decorator==5.1.1
distro==1.5.0
docutils==0.16
drumhat==0.1.0
elevenlabs==0.2.12
envirophat==1.0.0
executing==1.2.0
ExplorerHAT==0.4.2
filelock==3.12.0
Flask==1.1.2
fourletterphat==0.1.0
frozenlist==1.3.3
gpiozero==1.6.2
greenlet==2.0.2
html5lib==1.1
idna==2.10
ipython==8.13.2
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
langchain==0.0.167
lazy-object-proxy==0.0.0
llvmlite==0.40.0
logilab-common==1.8.1
lxml==4.6.3
MarkupSafe==1.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mccabe==0.6.1
microdotphat==0.2.1
more-itertools==9.1.0
mote==0.0.4
motephat==0.0.3
mpmath==1.3.0
multidict==6.0.4
mypy==0.812
mypy-extensions==0.4.3
networkx==3.1
numba==0.57.0
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.1.0
openai==0.27.6
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.72
packaging==23.1
pantilthat==0.0.7
parso==0.8.1
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera2==0.3.9
pickleshare==0.7.5
pidng==4.0.9
piexif==1.1.3
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
prompt-toolkit==3.0.38
psutil==5.8.0
pure-eval==0.2.2
pvporcupine==2.2.0
pycairo==1.16.2
pycparser==2.21
pycups==2.0.1
pydantic==1.10.7
pygame==1.9.6
Pygments==2.7.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5b0
pysmbc==1.0.23
python-apt==2.2.1
python-prctl==1.7
PyYAML==6.0
rainbowhat==0.1.0
regex==2023.5.5
reportlab==3.5.59
requests==2.30.0
requests-oauthlib==1.0.0
responses==0.12.1
roman==2.0.0
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scipy==1.10.1
scrollphat==0.0.7
scrollphathd==1.2.1
Send2Trash==1.6.0b1
sense-hat==2.4.0
simplejpeg==1.6.4
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.2.1
spidev==3.5
SQLAlchemy==2.0.13
ssh-import-id==5.10
stack-data==0.6.2
sympy==1.12
tenacity==8.2.2
thonny==4.0.1
tiktoken==0.3.3
toml==0.10.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
touchphat==0.0.1
tqdm==4.65.0
traitlets==5.9.0
twython==3.8.2
typed-ast==1.4.2
typing-extensions==4.5.0
typing-inspect==0.8.0
unicornhathd==0.0.4
urllib3==1.26.5
v4l2-python3==0.3.2
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.9.2
